Add server module
